feat: add Carousel and Opponent component
